Modified packaging structure.
