+++ test/CheckTest.java
@@ -1,12 +1,11 @@
 import Entities.King;
-import Other.Checkmate.Checkmate;
 import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
-import Other.GameState;
-import Other.LoadGame;
-import Other.Checkmate.Check;
+import Controllers.GameState;
+import Controllers.LoadGame;
+import UseCases.Checkmate.Check;
 
 
 /**
