+++ test/QueenMoveTest.java
@@ -1,11 +1,8 @@
-import org.junit.Before;
 import org.junit.Test;
 import static org.junit.Assert.*;
 
-import Entities.*;
-import Entities.ChessPiece;
-import Other.GameState;
-import Other.LoadGame;
+import Controllers.GameState;
+import Controllers.LoadGame;
 
 /**
  * This Class is responsible for testing the movements of our Queen chess
